requirements.txt
asgiref==3.3.1
Django==3.1.5
django-model-utils==4.1.1
djangorestframework==3.12.2
pytz==2020.5
simplejson==3.17.2
sqlparse==0.4.1
